formatting

+++ functions/controller.employeeFunction.js
@@ -5,22 +5,24 @@ const Cryptr = require("cryptr");
 const cryptr = new Cryptr("myTotalySecretKey");
 const jwt = require("jsonwebtoken");
 const sec = require("../config");
-const Verifier = require("email-verifier");
 
 class employee {
   constructor() {}
 
-  signUp(payload) {
-    return new Promise((resolve, reject) => {
-      let verifier = new Verifier("https://emailverification.whoisxmlapi.com/api/v1?apiKey=at_6yki4sbvWdEfP4Don22wi8Wot8WTQ&emailAddress=support@whoisxmlapi.com");
+  signUp(payload,isadmin) {
 
-     verifier.verify("mahat.ashin@gmail.com", (err, data) => {
-       console.log(payload.email)
-        if (err) {
-          console.log("error")
-        }
-        console.log(data);
-    });
+    
+    return new Promise((resolve, reject) => {
+      if(isadmin){
+    //   let verifier = new Verifier("https://emailverification.whoisxmlapi.com/api/v1?apiKey=at_6yki4sbvWdEfP4Don22wi8Wot8WTQ&emailAddress=support@whoisxmlapi.com");
+
+    //  verifier.verify("mahat.ashin@gmail.com", (err, data) => {
+    //    console.log(payload.email)
+    //     if (err) {
+    //       console.log("error")
+    //     }
+    //     console.log(data);
+    // });
       console.log("dada", payload.fullName);
       employeeContact
         .findOne({
@@ -31,6 +33,7 @@ class employee {
           if (user) {
             resolve("user already exist");
           } else {
+
             let obj = new employeeContact(payload);
             // obj.email=payload.email;
             // validator.validate(obj.email); // true
@@ -46,9 +49,16 @@ class employee {
             console.log("break");
           }
         });
-    });
+    }
+    else{
+      resolve("you are not admin")
+    }
+  });
+  
   }
-
+  
+  
+//login
   login(payload) {
     return new Promise((resolve, reject) => {
       console.log("adka ajdk");
@@ -96,6 +106,9 @@ class employee {
     });
   }
 
+
+  //get approve details
+
   getApproveDetails(userId)
   {
    const u=userId
@@ -113,15 +126,29 @@ class employee {
     });
   }
 
-  getAllEmployee() {
+
+  //get all employee
+  getAllEmployee(isadmin) {
+
+    
     return new Promise((resolve, reject) => {
+      if(isadmin)
+      {
+
       employeeContact
         .find()
         .then(d => resolve(d))
         .catch(e => reject(e));
-    });
+    }
+    else
+{
+    resolve("you are not an admin")
+}
+ });
   }
 
+
+  //find particular employee
   findParticular(userId) {
     return new Promise((resolve, reject) => {
       console.log("employee get", userId);
@@ -133,9 +160,15 @@ class employee {
     });
   }
 
-  update(userId, payload) {
+
+
+  //update employee details
+
+  update(userId,isadmin,Uid, payload) {
     return new Promise((resolve, reject) => {
-      console.log("ada");
+
+      if(isadmin || userId== Uid)
+      { console.log("ada");
       employeeContact
         .findOneAndUpdate(
           {
@@ -150,27 +183,54 @@ class employee {
         )
         .then(d => resolve(d))
         .catch(e => reject(e));
+      }
+     else
+     {
+       resolve("you are not an authorized")
+     }
     });
   }
  
-  removeEmployee(userId) {
+
+  //remove employee
+  removeEmployee(userId,isadmin) {
     return new Promise((resolve, reject) => {
-      employeeContact
-        .findByIdAndRemove(userId)
+      if(isadmin)
+      {
+        employeeContact
+        .findByIdAndDelete(userId)
         .then(d => resolve(d))
         .catch(e => reject(e));
-    });
-  }
-  removeEmployeeLeave(eid) {
-    return new Promise((resolve, reject) => {
-      employeeLeaveDetails
-        .deleteMany({eid:eid})
-        .then(d => resolve(d))
-        .catch(e => reject(e));
+      }
+      else
+      {
+        resolve("you are not admin")
+      }
+      
     });
   }
 
-  resetPassword(userId,email,Currentpassword,payload)
+  //remove employee leave 
+  removeEmployeeLeave(eid,isadmin) {
+    return new Promise((resolve, reject) => {
+      if(isadmin)
+      {
+        employeeLeaveDetails
+        .deleteMany({eid:eid})
+        .then(d => resolve(d))
+        .catch(e => reject(e));
+      }
+      else
+      {
+        resolve("you are not admin")
+      }
+    });
+  }
+
+
+  //reset password
+
+  resetPassword(userId,email,payload)
   {
